fix(RHINENG-19582): Remove old PDF generator
