+++ package.json
@@ -79,7 +79,6 @@
     "@redhat-cloud-services/frontend-components-advisor-components": "^2.0.16",
     "@redhat-cloud-services/frontend-components-config": "^6.5.2",
     "@redhat-cloud-services/frontend-components-notifications": "^4.1.17",
-    "@redhat-cloud-services/frontend-components-pdf-generator": "^4.0.7",
     "@redhat-cloud-services/frontend-components-remediations": "3.2.28",
     "@redhat-cloud-services/frontend-components-translations": "3.2.19",
     "@redhat-cloud-services/frontend-components-utilities": "^5.0.13",
