+++ src/PresentationalComponents/ExecutiveReport/Download.js
@@ -1,26 +1,20 @@
 import './_Download.scss';
+import React, { useContext, useState } from 'react';
 import PropTypes from 'prop-types';
-import {
-  INSIGHTS_HEADER,
-  RULES_FETCH_URL,
-  STATS_REPORTS_FETCH_URL,
-  STATS_SYSTEMS_FETCH_URL,
-  exportNotifications,
-} from '../../AppConstants';
-import React, { useState } from 'react';
-import { DownloadButton } from '@redhat-cloud-services/frontend-components-pdf-generator';
+import { useDispatch } from 'react-redux';
 import { addNotification } from '@redhat-cloud-services/frontend-components-notifications/redux/actions/notifications';
-import { Get } from '../../Utilities/Api';
+
+import { ExportIcon } from '@patternfly/react-icons';
 import messages from '../../Messages';
 import { useIntl } from 'react-intl';
-import { useDispatch } from 'react-redux';
-import { ExportIcon } from '@patternfly/react-icons';
-import { populateExportError } from '../helper';
-import BuildExecReport from './BuildExecReport';
+import { Button } from '@patternfly/react-core';
+import { exportNotifications } from '../../AppConstants';
+import { EnvironmentContext } from '../../App';
 
-const DownloadExecReport = ({ isDisabled }) => {
+const NewDownloadExecReport = ({ isDisabled }) => {
   const intl = useIntl();
   const dispatch = useDispatch();
+  const envContext = useContext(EnvironmentContext);
   const [loading, setLoading] = useState(false);
 
   const dataFetch = async () => {
@@ -28,68 +22,48 @@ const DownloadExecReport = ({ isDisabled }) => {
     dispatch(addNotification(exportNotifications.pending));
 
     try {
-      const [statsSystems, statsReports, topActiveRec] = (
-        await Promise.all([
-          Get(STATS_SYSTEMS_FETCH_URL),
-          Get(STATS_REPORTS_FETCH_URL),
-          Get(
-            RULES_FETCH_URL,
-            {},
-            { limit: 3, sort: '-total_risk,-impacted_count', impacting: true },
-          ),
-        ])
-      ).map(({ data }) => data);
-
-      const report = (
-        <BuildExecReport
-          statsReports={statsReports}
-          statsSystems={statsSystems}
-          topActiveRec={topActiveRec}
-          intl={intl}
-        />
-      );
+      await envContext.requestPdf({
+        filename: `Advisor-Executive-Report--${new Date()
+          .toUTCString()
+          .replace(/ /g, '-')}.pdf`,
+        payload: {
+          manifestLocation: '/apps/advisor/fed-mods.json',
+          scope: 'advisor',
+          module: './NewBuildExecReport',
+          fetchDataParams: {
+            limit: 3,
+            sort: '-total_risk,-impacted_count',
+            impacting: true,
+          },
+        },
+      });
 
       setLoading(false);
-
       dispatch(addNotification(exportNotifications.success));
-
-      return [report];
-    } catch (e) {
+    } catch (error) {
+      void error;
       setLoading(false);
-      dispatch(addNotification(populateExportError(e)));
-
-      return [];
+      dispatch(addNotification(exportNotifications.error));
     }
   };
 
   return (
-    <DownloadButton
-      fallback={<div />}
-      groupName={intl.formatMessage(messages.redHatInsights)}
-      label={
-        loading
-          ? intl.formatMessage(messages.loading)
-          : intl.formatMessage(messages.downloadExecutiveLabel)
-      }
-      asyncFunction={dataFetch}
-      buttonProps={{
-        variant: 'link',
-        icon: <ExportIcon className="iconOverride" />,
-        component: 'a',
-        className: 'downloadButtonOverride',
-        isAriaDisabled: isDisabled,
-        ...(loading ? { isDisabled: true } : null),
-      }}
-      type={INSIGHTS_HEADER}
-      fileName={`Advisor-Executive-Report--${new Date()
-        .toUTCString()
-        .replace(/ /g, '-')}.pdf`}
-    />
+    <Button
+      onClick={dataFetch}
+      variant="link"
+      isInline
+      isDisabled={isDisabled || loading}
+      icon={<ExportIcon className="iconOverride" />}
+      className="downloadButtonOverride"
+      aria-label="Download Exec Report"
+    >
+      {intl.formatMessage(messages.downloadExecutiveLabel)}
+    </Button>
   );
 };
 
-DownloadExecReport.propTypes = {
+NewDownloadExecReport.propTypes = {
   isDisabled: PropTypes.bool,
 };
 
-export default DownloadExecReport;
+export default NewDownloadExecReport;
